Revert "Task 75, Assignment 1, 2 added"
