+++ src/Lesson2/Task66/Assignment2/com/fruitbase/FruitBase.java
@@ -1,5 +1,4 @@
 package Lesson2.Task66.Assignment2.com.fruitbase;
-
 import Lesson2.Task66.Assignment2.com.fruitbase.fruits.Fruit;
 
 public class FruitBase {
@@ -21,4 +20,15 @@ public class FruitBase {
         return cargo;
     }
 
+    public static void main(String[] args) {
+        FruitBase base = new FruitBase();
+        if (args.length == 0) {
+            System.out.println("Заказ не указан в аргументах командной строки");
+            return;
+        }
+        Cargo cargo = base.takeOrder(args);
+        System.out.println("Информация о грузе \n" +
+                "Вес: " + cargo.getWeight() + "\n"
+                + "Цена: " + cargo.getPrice());
+    }
 }
