+++ src/Lesson2/Task66/Assignment2/com/fruitbase/fruits/Apple.java
@@ -3,7 +3,7 @@ package Lesson2.Task66.Assignment2.com.fruitbase.fruits;
 import java.math.BigDecimal;
 
 public class Apple extends Fruit {
-    public Apple(String name, double weight, BigDecimal price) {
-        super(name, weight, price);
+    public Apple(double weight, BigDecimal price, String name){
+        super(weight, price, name);
     }
 }
