+++ src/Lesson2/Task66/Assignment2/com/fruitbase/fruits/Banana.java
@@ -2,8 +2,8 @@ package Lesson2.Task66.Assignment2.com.fruitbase.fruits;
 
 import java.math.BigDecimal;
 
-public class Banana extends Fruit {
-    public Banana(String name, double weight, BigDecimal price) {
-        super(name, weight, price);
+public class Banana extends Fruit{
+    public Banana(double weight, BigDecimal price, String name){
+        super(weight, price, name);
     }
 }
