--- src/Lesson2/Task66/Assignment2/com/fruitbase/fruits/Freshness.java
@@ -1,7 +1,0 @@
-package Lesson2.Task66.Assignment2.com.fruitbase.fruits;
-
-public enum Freshness {
-    FRESH,
-    OVERRIPED,
-    SPOILED
-}
