+++ src/Lesson2/Task66/Assignment2/com/fruitbase/fruits/Fruit.java
@@ -2,19 +2,15 @@ package Lesson2.Task66.Assignment2.com.fruitbase.fruits;
 
 import java.math.BigDecimal;
 
-import static Lesson2.Task66.Assignment2.com.fruitbase.fruits.Freshness.FRESH;
-
 public abstract class Fruit {
-    private String name;
     private double weight;
     private BigDecimal price;
-    private Freshness freshness;
+    private String name;
 
-    public Fruit(String name, double weight, BigDecimal price) {
-        this.name = name;
+    public Fruit(double weight, BigDecimal price, String name) {
         this.weight = weight;
         this.price = price;
-        this.freshness = FRESH;
+        this.name = name;
     }
 
     public double getWeight() {
@@ -28,34 +24,4 @@ public abstract class Fruit {
     public String getName() {
         return name;
     }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj == this) {
-            return true;
-        }
-        if (obj == null || obj.getClass() != this.getClass()) {
-            return false;
-        }
-
-        Fruit fruit = (Fruit) obj;
-        return name == fruit.name
-                && (weight == fruit.weight) && (price == fruit.price);
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((name == null) ? 0 : name.hashCode());
-        result = prime * result + (int) weight * 1000;
-        result = prime * result + price.hashCode();
-        result = prime * result + freshness.hashCode();
-        return result;
-    }
-
-    public boolean isFresh() {
-        return freshness == FRESH;
-    }
-
 }
