+++ src/Lesson2/Task66/Assignment2/com/fruitbase/fruits/Orange.java
@@ -3,7 +3,7 @@ package Lesson2.Task66.Assignment2.com.fruitbase.fruits;
 import java.math.BigDecimal;
 
 public class Orange extends Fruit {
-    public Orange(String name, double weight, BigDecimal price) {
-        super(name, weight, price);
+    public Orange(double weight, BigDecimal price, String name){
+        super(weight, price, name);
     }
 }
