+++ src/Lesson2/Task66/Assignment2/com/fruitbase/fruits/Pineapple.java
@@ -3,7 +3,7 @@ package Lesson2.Task66.Assignment2.com.fruitbase.fruits;
 import java.math.BigDecimal;
 
 public class Pineapple extends Fruit {
-    public Pineapple(String name, double weight, BigDecimal price) {
-        super(name, weight, price);
+    public  Pineapple(double weight, BigDecimal price, String name){
+        super(weight, price, name);
     }
 }
